refactor: transaction.py position
